Dry run migrations
Skip first time setup and 6.5 migrations

+++ src/core/server/saved_objects/migrations/core/index_migrator.ts
@@ -22,6 +22,7 @@ import * as Index from './elastic_index';
 import { migrateRawDocs } from './migrate_raw_docs';
 import { Context, migrationContext, MigrationOpts } from './migration_context';
 import { coordinateMigration, MigrationResult } from './migration_coordinator';
+import { fetchInfo } from './elastic_index';
 
 /*
  * Core logic for migrating the mappings and documents in an index.
@@ -45,23 +46,65 @@ export class IndexMigrator {
    * @returns {Promise<MigrationResult>}
    */
   public async migrate(): Promise<MigrationResult> {
-    const context = await migrationContext(this.opts);
+    const { dryRun, index, callCluster, log, batchSize } = this.opts;
+    if (dryRun) {
+      const indexInfo = await fetchInfo(callCluster, index);
+      if (indexInfo.exists === false) {
+        return {
+          status: 'skipped',
+          alias: this.opts.index,
+          reason: `nothing to migrate, index ${index} doesn't exist.`,
+        };
+      } else if (!indexInfo.aliases[index]) {
+        return {
+          status: 'skipped',
+          alias: index,
+          reason: `expected an alias but found an index: ${index}.`,
+        };
+      } else if (await requiresMigration(await migrationContext(this.opts))) {
+        const dryRunAlias = index + '_dry_run';
+        const dryRunIndex = dryRunAlias + '_1';
+        log.info(
+          `Reindexing ${indexInfo.indexName} into ${dryRunIndex} for performing dry run migration.`
+        );
+        await Index.createIndex(callCluster, dryRunIndex, indexInfo.mappings, dryRunAlias);
+        await Index.reindex(callCluster, indexInfo.indexName, dryRunIndex, batchSize);
+      } else {
+        return { status: 'skipped', alias: index };
+      }
+    }
+
+    const context = await migrationContext({
+      ...this.opts,
+      index: this.opts.index + (dryRun ? '_dry_run' : ''),
+    });
 
     return coordinateMigration({
-      log: context.log,
-
+      alias: context.alias,
       pollInterval: context.pollInterval,
-
+      log: context.log,
       async isMigrated() {
         return !(await requiresMigration(context));
       },
 
       async runMigration() {
         if (await requiresMigration(context)) {
-          return migrateIndex(context);
+          const result = await migrateIndex(context, dryRun);
+          if (dryRun) {
+            context.log.info(
+              `Cleaning up after dry run migration by deleting: ${context.source.indexName}, ${context.dest.indexName}`
+            );
+            try {
+              await Index.deleteIndex(context.callCluster, context.dest.indexName);
+              await Index.deleteIndex(context.callCluster, context.source.indexName);
+            } catch (e) {
+              /* ignore*/
+            }
+          }
+          return result;
         }
 
-        return { status: 'skipped' };
+        return { status: 'skipped', alias: index };
       },
     });
   }
@@ -95,7 +138,9 @@ async function requiresMigration(context: Context): Promise<boolean> {
   const diffResult = diffMappings(refreshedSource.mappings, dest.mappings);
 
   if (diffResult) {
-    log.info(`Detected mapping change in "${diffResult.changedProp}"`);
+    log.info(
+      `Detected mapping change between in ${dest.indexName} and ${refreshedSource.indexName} for '${diffResult.changedProp}'`
+    );
 
     return true;
   }
@@ -107,11 +152,13 @@ async function requiresMigration(context: Context): Promise<boolean> {
  * Performs an index migration if the source index exists, otherwise
  * this simply creates the dest index with the proper mappings.
  */
-async function migrateIndex(context: Context): Promise<MigrationResult> {
+async function migrateIndex(context: Context, dryRun: boolean): Promise<MigrationResult> {
   const startTime = Date.now();
   const { callCluster, alias, source, dest, log } = context;
 
-  await deleteIndexTemplates(context);
+  if (!dryRun) {
+    await deleteIndexTemplates(context);
+  }
 
   log.info(`Creating index ${dest.indexName}.`);
 
@@ -120,17 +167,17 @@ async function migrateIndex(context: Context): Promise<MigrationResult> {
   await migrateSourceToDest(context);
 
   log.info(`Pointing alias ${alias} to ${dest.indexName}.`);
-
   await Index.claimAlias(callCluster, dest.indexName, alias);
 
   const result: MigrationResult = {
+    alias,
     status: 'migrated',
     destIndex: dest.indexName,
     sourceIndex: source.indexName,
     elapsedMs: Date.now() - startTime,
   };
 
-  log.info(`Finished in ${result.elapsedMs}ms.`);
+  log.info(`Finished migrating '${context.alias}' in ${result.elapsedMs}ms.`);
 
   return result;
 }
